alternative OneCallHistory API loader and a test.
Review code comments for calculator.py and calculator_utils.py .The rest of the project uses dataframes, so this will be more in line with the rest of the project and maybe easier to migrate later. This way we save all details provided by the API, so no need to call multiple times for the same timestamp.
big_df.csv is a result of calculator.py